I51 Rename ProjectTable to CompletableTable
