+++ client/src/main/components/CompletableTable/CompletableRow/NoteInput.tsx
