+++ client/src/main/components/CompletableTable/CompletableRow/SimpleTextInput.tsx
