+++ client/src/main/components/CompletableTable/CompletableRow/TagRow/index.tsx
