+++ client/src/main/components/CompletableTable/CompletableRow/TaskExpanderButton.tsx
