+++ client/src/main/components/CompletableTable/CompletableRow/index.tsx
