+++ client/src/main/components/CompletableTable/HiddenItemsCaption.tsx
