+++ client/src/main/components/CompletableTable/index.tsx
@@ -50,20 +50,18 @@ function styles(theme: Theme) {
   });
 }
 
-export interface ProjectTableProps extends WithStyles<typeof styles> {
+export interface CompletableTableProps extends WithStyles<typeof styles> {
   rootCompletableType?: CompletableType;
   rootCompletableId?: string;
 }
-
-export type ProjectTableState = unknown;
 
 /**
  * Represents the complete table of projects in the UI, as well as modification
  * components such as sorting and project addition buttons.
  *
- * @param {ProjectTableProps} props the props
+ * @param {CompletableTableProps} props the props
  */
-function ProjectTable(props: ProjectTableProps) {
+function CompletableTable(props: CompletableTableProps) {
   const { rootCompletableId, rootCompletableType, classes } = props;
 
   // The type of completables in the table
@@ -339,4 +337,4 @@ function ProjectTable(props: ProjectTableProps) {
   );
 }
 
-export default withStyles(styles, { withTheme: true })(ProjectTable);
+export default withStyles(styles, { withTheme: true })(CompletableTable);
